Let MatchPrefixResult return more than just the match on success.
This lets us return a context rather than modifying the one passed in
which gives the parent matcher the information it needs, since it
can't know whether the inner matcher is going to return an object or
a value; it might be an Alt.

This is a major refactoring and I used some rugs to do it yay me.